chore(core): rename libs
