--- libs/lexer/README.md
@@ -1,11 +1,0 @@
-# lexer
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build lexer` to build the library.
-
-## Running unit tests
-
-Run `nx test lexer` to execute the unit tests via [Vitest](https://vitest.dev/).
