--- libs/lexer/src/index.ts
@@ -1,8 +1,0 @@
-export type {
-  AttrToken,
-  ElementToken,
-  TextToken,
-  NodeToken,
-} from './lib/tokens.js';
-
-export { tokenize } from './lib/lexer.js';
