--- libs/lexer/src/lib/tokens.ts
@@ -1,23 +1,0 @@
-export type AttrToken = {
-  name: string;
-  value: string;
-};
-
-export type NodeToken = TextToken | ElementToken | CommentToken;
-
-export type TextToken = {
-  type: 'text';
-  content: string;
-};
-
-export type ElementToken = {
-  type: 'element';
-  name: keyof HTMLElementTagNameMap | (string & {});
-  attributes: AttrToken[];
-  children: NodeToken[];
-};
-
-export type CommentToken = {
-  type: 'comment';
-  content: string;
-};
