+++ libs/parser/package.json
@@ -16,12 +16,12 @@
     }
   },
   "dependencies": {
-    "tslib": "^2.3.0",
-    "@sievert/lexer": "workspace:*",
-    "@sievert/graph": "workspace:*",
-    "@sievert/signals": "workspace:*"
+    "tslib": "^2.3.0"
   },
   "nx": {
-    "tags": ["scope:core", "type:internal"]
+    "tags": [
+      "scope:core",
+      "type:internal"
+    ]
   }
 }
