+++ libs/parser/src/index.ts
@@ -1,1 +1,3 @@
-export { html } from './lib/html.js';
+export type { Attribute, ElementNode, TextNode, SvNode } from './lib/nodes.js';
+
+export { parse } from './lib/parser.js';
