+++ libs/parser/src/lib/parser.spec.ts
@@ -1,125 +1,278 @@
-import { Params } from './params.js';
 import { parse } from './parser.js';
-
-const emptyParams = (): Params => ({
-  keys: [],
-  expressions: [],
-});
+import { SvNode } from './nodes.js';
 
 describe('parser', () => {
-  it('should return empty result with empty input', () => {
-    const result = parse({
-      tokens: [],
-      params: emptyParams(),
+  describe('elements', () => {
+    it('should parse element', () => {
+      const tokens = parse('<test></test>');
+
+      expect(tokens).toEqual<SvNode[]>([
+        { type: 'element', name: 'test', attributes: [], children: [] },
+      ]);
     });
 
-    expect(result.documentFragment.children.length).toBe(0);
-    expect(result.sinks.length).toBe(0);
-  });
+    it('should parse nested elements', () => {
+      const tokens = parse('<test><test></test></test>');
 
-  it('should parse elements', () => {
-    const result = parse({
-      tokens: [{ type: 'element', name: 'div', attributes: [], children: [] }],
-      params: emptyParams(),
-    });
-
-    expect(result.documentFragment.children.length).toBe(1);
-    expect(result.documentFragment.children[0]).toBeInstanceOf(HTMLDivElement);
-  });
-
-  it('should parse attributes', () => {
-    const result = parse({
-      tokens: [
+      expect(tokens).toEqual<SvNode[]>([
         {
           type: 'element',
-          name: 'div',
-          attributes: [{ name: 'test', value: 'sievert' }],
+          name: 'test',
+          attributes: [],
+          children: [
+            { type: 'element', name: 'test', attributes: [], children: [] },
+          ],
+        },
+      ]);
+    });
+
+    it('should parse void element', () => {
+      const tokens = parse('<test />');
+
+      expect(tokens).toEqual<SvNode[]>([
+        { type: 'element', name: 'test', attributes: [], children: [] },
+      ]);
+    });
+
+    it('should not expect closing tag for void element', () => {
+      expect(() => parse('<test/></test>')).toThrow(
+        'Unexpected closing tag at position 7',
+      );
+    });
+
+    it('should not parse elements following void elements as children', () => {
+      const tokens = parse('<test/><test></test>');
+
+      expect(tokens).toEqual<SvNode[]>([
+        { type: 'element', name: 'test', attributes: [], children: [] },
+        { type: 'element', name: 'test', attributes: [], children: [] },
+      ]);
+    });
+
+    it('should not parse text following void elements as children', () => {
+      const tokens = parse('<test/>test');
+
+      expect(tokens).toEqual<SvNode[]>([
+        { type: 'element', name: 'test', attributes: [], children: [] },
+        { type: 'text', content: 'test' },
+      ]);
+    });
+
+    it('should parse names with letters, numbers and hyphens', () => {
+      const tokens = parse('<test-123></test-123>');
+
+      expect(tokens).toEqual<SvNode[]>([
+        { type: 'element', name: 'test-123', attributes: [], children: [] },
+      ]);
+    });
+
+    it('should throw on unterminated opening tag', () => {
+      expect(() => parse('<test')).toThrow(
+        'Unterminated opening tag with name "test" at position 5',
+      );
+    });
+
+    it('should throw on invalid element name', () => {
+      expect(() => parse('<test$></test$>')).toThrow(
+        `Unexpected terminator "$" for tag with name "test" at position 5`,
+      );
+    });
+
+    it('should throw on unexpected closing tag', () => {
+      expect(() => parse('<test-1></test-2>')).toThrow(
+        'Unexpected closing tag name "test-2" for tag with name "test-1" at position 10',
+      );
+    });
+
+    it('should throw on unterminated closing tag', () => {
+      expect(() => parse('<test></test')).toThrow(
+        'Unexpected terminator "undefined" for closing tag with name "test" at position 12',
+      );
+    });
+  });
+
+  describe('attributes', () => {
+    it('should parse bool attribute', () => {
+      const tokens = parse('<test test></test>');
+
+      expect(tokens).toEqual<SvNode[]>([
+        {
+          type: 'element',
+          name: 'test',
+          attributes: [{ name: 'test', value: '' }],
           children: [],
         },
-      ],
-      params: emptyParams(),
+      ]);
     });
 
-    const div = result.documentFragment.children[0];
-    expect(div.getAttribute('test')).toBe('sievert');
-  });
+    it('should parse quoted and unquoted attributes', () => {
+      const tokens = parse(
+        '<test double="value" single=\'value\' none=value></test>',
+      );
 
-  it('should parse text', () => {
-    const result = parse({
-      tokens: [
+      expect(tokens).toEqual<SvNode[]>([
         {
           type: 'element',
-          name: 'div',
+          name: 'test',
+          attributes: [
+            { name: 'double', value: 'value' },
+            { name: 'single', value: 'value' },
+            { name: 'none', value: 'value' },
+          ],
+          children: [],
+        },
+      ]);
+    });
+
+    it('should throw on invalid attribute name', () => {
+      expect(() => parse('<test $test></test>')).toThrow(
+        'Unexpected terminator "$" for tag with name "test" at position 6',
+      );
+    });
+
+    it('should throw on unterminated attribute', () => {
+      expect(() => parse('<test test="test')).toThrow(
+        'Unterminated attribute with name "test" at position 16',
+      );
+    });
+  });
+
+  describe('comments', () => {
+    it('should parse comments', () => {
+      const tokens = parse('<!--test-->');
+
+      expect(tokens).toEqual<SvNode[]>([
+        {
+          type: 'comment',
+          content: 'test',
+        },
+      ]);
+    });
+
+    it('should throw on bogus comment', () => {
+      expect(() => parse('<!test>')).toThrow(
+        'Unexpected bogus comment at position 3',
+      );
+    });
+
+    it('should throw on unterminated comment', () => {
+      expect(() => parse('<!--test')).toThrow(
+        'Unexpected terminator for comment at position 8',
+      );
+    });
+  });
+
+  describe('text', () => {
+    it('should parse text', () => {
+      const tokens = parse('test');
+
+      expect(tokens).toEqual<SvNode[]>([{ type: 'text', content: 'test' }]);
+    });
+
+    it('should flush text buffer before element', () => {
+      const tokens = parse('test<test></test>');
+
+      expect(tokens).toEqual<SvNode[]>([
+        { type: 'text', content: 'test' },
+        { type: 'element', name: 'test', attributes: [], children: [] },
+      ]);
+    });
+
+    it('should flush text buffer before closing element', () => {
+      const tokens = parse('<test>test</test>');
+
+      expect(tokens).toEqual<SvNode[]>([
+        {
+          type: 'element',
+          name: 'test',
           attributes: [],
           children: [{ type: 'text', content: 'test' }],
         },
-      ],
-      params: emptyParams(),
+      ]);
     });
 
-    const div = result.documentFragment.children[0];
-    expect(div.childNodes.length).toBe(1);
-    expect(div.childNodes[0].textContent).toBe('test');
+    it('should flush text buffer before comment', () => {
+      const tokens = parse('test<!--test-->');
+
+      expect(tokens).toEqual<SvNode[]>([
+        { type: 'text', content: 'test' },
+        { type: 'comment', content: 'test' },
+      ]);
+    });
   });
 
-  it('should ignore unused params', () => {
-    const result = parse({
-      tokens: [],
-      params: {
-        keys: ['key'],
-        expressions: ['test'],
+  it('should parse mixed content', () => {
+    const tokens = parse(`
+      <div id="test">
+        <a href="#">Link</a>
+        Hello
+        <span>World</span>
+        <img alt='test' src=test />
+        <!-- comment -->
+      </div>
+    `);
+
+    expect(tokens).toEqual<SvNode[]>([
+      {
+        type: 'text',
+        content: '\n      ',
       },
-    });
-
-    expect(result.documentFragment.children.length).toBe(0);
-    expect(result.sinks.length).toBe(0);
-  });
-
-  it('should take param matches', () => {
-    const params = {
-      keys: ['key'],
-      expressions: ['value'],
-    };
-
-    const result = parse({
-      tokens: [
-        {
-          type: 'element',
-          name: 'div',
-          attributes: [],
-          children: [{ type: 'text', content: 'key' }],
-        },
-      ],
-      params,
-    });
-
-    expect(params.keys.length).toBe(0);
-    expect(params.expressions.length).toBe(0);
-    expect(result.sinks.length).toBe(0);
-  });
-
-  it('should parse event listeners', () => {
-    const cb = vi.fn();
-
-    const result = parse({
-      tokens: [
-        {
-          type: 'element',
-          name: 'button',
-          attributes: [{ name: 'onclick', value: 'key' }],
-          children: [],
-        },
-      ],
-      params: {
-        keys: ['key'],
-        expressions: [cb],
+      {
+        type: 'element',
+        name: 'div',
+        attributes: [{ name: 'id', value: 'test' }],
+        children: [
+          {
+            type: 'text',
+            content: '\n        ',
+          },
+          {
+            type: 'element',
+            name: 'a',
+            attributes: [{ name: 'href', value: '#' }],
+            children: [{ type: 'text', content: 'Link' }],
+          },
+          {
+            type: 'text',
+            content: '\n        Hello\n        ',
+          },
+          {
+            type: 'element',
+            name: 'span',
+            attributes: [],
+            children: [{ type: 'text', content: 'World' }],
+          },
+          {
+            type: 'text',
+            content: '\n        ',
+          },
+          {
+            type: 'element',
+            name: 'img',
+            attributes: [
+              { name: 'alt', value: 'test' },
+              { name: 'src', value: 'test' },
+            ],
+            children: [],
+          },
+          {
+            type: 'text',
+            content: '\n        ',
+          },
+          {
+            type: 'comment',
+            content: ' comment ',
+          },
+          {
+            type: 'text',
+            content: '\n      ',
+          },
+        ],
       },
-    });
-
-    const button = result.documentFragment.children[0];
-    expect(cb).not.toHaveBeenCalled();
-
-    button.dispatchEvent(new PointerEvent('click'));
-    expect(cb).toHaveBeenCalledOnce();
+      {
+        type: 'text',
+        content: '\n    ',
+      },
+    ]);
   });
 });
