+++ libs/parser/tsconfig.lib.json
@@ -10,17 +10,7 @@
     "types": ["node"]
   },
   "include": ["src/**/*.ts"],
-  "references": [
-    {
-      "path": "../signals/tsconfig.lib.json"
-    },
-    {
-      "path": "../graph/tsconfig.lib.json"
-    },
-    {
-      "path": "../lexer/tsconfig.lib.json"
-    }
-  ],
+  "references": [],
   "exclude": [
     "vite.config.ts",
     "vite.config.mts",
