+++ libs/renderer/README.md
@@ -1,0 +1,11 @@
+# renderer
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build renderer` to build the library.
+
+## Running unit tests
+
+Run `nx test renderer` to execute the unit tests via [Vitest](https://vitest.dev/).
