+++ libs/renderer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@sievert/lexer",
+  "name": "@sievert/renderer",
   "version": "0.0.1",
   "private": true,
   "type": "module",
@@ -16,9 +16,15 @@
     }
   },
   "dependencies": {
-    "tslib": "^2.3.0"
+    "tslib": "^2.3.0",
+    "@sievert/parser": "workspace:*",
+    "@sievert/graph": "workspace:*",
+    "@sievert/signals": "workspace:*"
   },
   "nx": {
-    "tags": ["scope:core", "type:internal"]
+    "tags": [
+      "scope:core",
+      "type:internal"
+    ]
   }
 }
