+++ libs/renderer/src/index.ts
@@ -1,0 +1,1 @@
+export { html } from './lib/html.js';
