+++ tsconfig.base.json
@@ -21,7 +21,7 @@
     "rootDir": ".",
     "paths": {
       "@sievertjs/core": ["packages/core/src/index.ts"],
-      "@sievert/lexer": ["libs/lexer/src/index.ts"],
+      "@sievert/parser": ["libs/parser/src/index.ts"],
       "@sievert/graph": ["libs/graph/src/index.ts"],
       "@sievert/signals": ["libs/signals/src/index.ts"]
     }
